+++ src/AudioPlayer/test/AudioPlayer.e2eStory.js
@@ -60,7 +60,8 @@ storiesOf(kind, module).add(testStories.audioPlayerLazyLoad, () => {
     <AudioPlayer
       dataHook={storySettings.dataHookLazyLoad}
       src="https://music.wixstatic.com/preview/84770f_29a80480c24f4946a477c8ad03b92cbc-128.mp3"
-      lazyLoad
+      preload="none"
+      webAudioAPI
     />
   );
 });