+++ CHANGELOG.md
@@ -28,6 +28,10 @@ Types of changes:
 - `Input` & `Selection` families- changing large size [#5457](https://github.com/wix/wix-style-react/pull/5457)
 - `<Notification/>`- migrating styles to stylable and updating drivers accordingly [#5484](https://github.com/wix/wix-style-react/pull/5484)
 
+### Breaking
+- `<AudioPlayer>` - Renamed prop `html5Audio` to `webAudioAPI` which makes the player to be html5 by default [#5488](https://github.com/wix/wix-style-react/pull/5488)
+- `<AudioPlayer>` - Changed prop `lazyLoad` (boolean) to be `preload` ('auto', 'metadata' or 'none') defaults to 'metadata'. (lazyLoad=false equals to prelaod=auto) [#5488](https://github.com/wix/wix-style-react/pull/5488)
+
 ## 8.11.0 - 2020-05-19
 ### Added
 - `<SortableGrid/>` - new component [#5467](https://github.com/wix/wix-style-react/pull/5467)